backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.10.0
    # via
    #   -r requirements.in
    #   django
    #   django-cors-headers
black==25.9.0
    # via -r requirements.in
build==1.3.0
    # via
    #   -r requirements.in
    #   pip-tools
click==8.3.0
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
django==5.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   -r requirements.in
    #   black
packaging==25.0
    # via
    #   -r requirements.in
    #   black
    #   build
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pip-tools==7.5.1
    # via -r requirements.in
platformdirs==4.5.0
    # via
    #   -r requirements.in
    #   black
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
pytokens==0.2.0
    # via
    #   -r requirements.in
    #   black
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
tzdata==2025.2
    # via
    #   -r requirements.in
    #   django
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
